interface segregation based on Accedo One endpoints
