+++ .gitignore
@@ -1,1 +1,5 @@
+AccedoOneiOS/AccedoOneiOS.xcodeproj/project.xcworkspace/xcuserdata/stephen.xcuserdatad/
+AccedoOneiOS/AccedoOneiOS.xcodeproj/xcuserdata/stephen.xcuserdatad/
+AccedoOnetvOS/AccedoOnetvOS.xcodeproj/project.xcworkspace/xcuserdata/stephen.xcuserdatad/
+AccedoOnetvOS/AccedoOnetvOS.xcodeproj/xcuserdata/stephen.xcuserdatad/
 /Binary/*
