+++ Common/API/AccedoOneUserData.m
@@ -13,9 +13,10 @@
 static NSString *const kPathUserDataApp      = @"user";
 static NSString *const kPathUserDataAppGroup = @"group";
 
-@interface AccedoOneUserData ()
+@interface AccedoOneUserData()
 @property (nonatomic, strong) AccedoOne * service;
 @end
+
 
 @implementation AccedoOneUserData
 
@@ -26,18 +27,17 @@ static NSString *const kPathUserDataAppGroup = @"group";
     return self;
 }
 
+#pragma mark - AccedoOneUserDataProtocol
+
 - (void) allDataForUser:(NSString *)userId scope:(AOUserDataScope)scope onComplete:(void (^)(NSDictionary *userData, AOError *err))completionBlock {
     
     NSString *requestPath = [[self pathForScope:scope] stringByAppendingPathComponent: userId];
     
-    [self.service sendAuthenticatedGETRequest:requestPath queryParams:nil allowCache:NO onSuccess:^(NSDictionary *response)
-     {
+    [self.service sendAuthenticatedGETRequest:requestPath queryParams:nil allowCache:NO onSuccess:^(NSDictionary *response) {
          if (completionBlock) {
              completionBlock(response, nil);
          }
-     }
-                         failureBlock:^(AOError *error)
-     {
+    } failureBlock:^(AOError *error) {
          if (completionBlock) {
              completionBlock(nil, error);
          }
@@ -45,21 +45,17 @@ static NSString *const kPathUserDataAppGroup = @"group";
 }
 
 - (void) storeData:(NSDictionary *)data forUser:(NSString *)userId scope:(AOUserDataScope)scope onComplete:(void(^)(AOError *err))completionBlock {
-    
+
     NSString *requestPath = [[self pathForScope:scope] stringByAppendingPathComponent: userId];
-    
     NSData * body = [NSJSONSerialization dataWithJSONObject:data options:0 error:nil];
-    
+
     [self.service setHeaderValue:@"application/json; charset=utf-8" headerField:@"Content-Type"];
     
-    [self.service sendAuthenticatedPOSTRequest:requestPath params:nil body:body onSuccess:^(id response)
-     {
+    [self.service sendAuthenticatedPOSTRequest:requestPath params:nil body:body onSuccess:^(id response) {
          if (completionBlock) {
              completionBlock(nil);
          }
-     }
-                             onFailure:^(AOError *error)
-     {
+    } onFailure:^(AOError *error) {
          if (completionBlock) {
              completionBlock(error);
          }
@@ -67,25 +63,19 @@ static NSString *const kPathUserDataAppGroup = @"group";
 }
 
 - (void) dataForUser:(NSString *)userId key:(NSString *)key scope:(AOUserDataScope)scope onComplete:(void(^)(NSString *value, AOError *err))completionBlock {
-    
+
     NSString *requestPath = [NSString pathWithComponents: @[[self pathForScope:scope], userId, key]];
     
-    [self.service sendAuthenticatedGETRequest:requestPath queryParams:nil allowCache:NO onSuccess:^(id response)
-     {
+    [self.service sendAuthenticatedGETRequest:requestPath queryParams:nil allowCache:NO onSuccess:^(id response) {
          if (completionBlock) {
              completionBlock(response, nil);
          }
-     }
-                         failureBlock:^(AOError *error)
-     {
-         if (error.code == 200 && error.responseObject) //hack solution, as response is not a JSON but a plain string...
-         {
+     } failureBlock:^(AOError *error) {
+         if (error.code == 200 && error.responseObject) { //hack solution, as response is not a JSON but a plain string...
              if (completionBlock) {
                  completionBlock(error.responseObject, nil);
              }
-         }
-         else
-         {
+         } else {
              if (completionBlock) {
                  completionBlock(nil, error);
              }
@@ -100,14 +90,11 @@ static NSString *const kPathUserDataAppGroup = @"group";
     
     [self.service setHeaderValue:@"text/plain; charset=utf-8" headerField:@"Content-Type"];
     
-    [self.service sendAuthenticatedPOSTRequest:requestPath params:nil body:body onSuccess:^(id response)
-     {
+    [self.service sendAuthenticatedPOSTRequest:requestPath params:nil body:body onSuccess:^(id response) {
          if (completionBlock) {
              completionBlock(nil);
          }
-     }
-                             onFailure:^(AOError *error)
-     {
+    } onFailure:^(AOError *error) {
          if (completionBlock) {
              completionBlock(error);
          }
@@ -118,12 +105,10 @@ static NSString *const kPathUserDataAppGroup = @"group";
 
 - (NSString *) pathForScope:(AOUserDataScope)scope {
     switch (scope) {
-        case AOUserDataScopeApplication:      return kPathUserDataApp;
+        case AOUserDataScopeApplication: return kPathUserDataApp;
         case AOUserDataScopeApplicationGroup: return kPathUserDataAppGroup;
         default: return nil;
     }
 }
-
-
 
 @end
